requirements.txt
# FastAPI 核心框架
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Google Gemini API
google-generativeai>=0.3.0

# 环境变量管理
python-dotenv>=1.0.0

# 测试工具
requests>=2.28.0
